Extract routing hook and feedback detail subcomponents from agents-view
Split agents-view.tsx (904 → 774 lines) by extracting:
- use-agents-view-routing.ts: route matching, feedback detail state
  derivation, route validation effects, and navigation callbacks
- agents-view-feedback-detail.tsx: DesktopFeedbackDetail and
  MobileFeedbackDetail subcomponents replacing inline IIFEs
